fix: prev next invalid
